list of changes
1. bitmap images did not display, issues was lv_draw_SW_support_(color
format) in lv_conf file.
2. Fixed a few issues with old project files casussing inclusion of FATS
library where it was not enabled.

+++ H743_DASH_LOGGER/Drivers/lvgl/src/misc/lv_utils.h
@@ -14,7 +14,8 @@ extern "C" {
  *      INCLUDES
  *********************/
 
-#include "../lvgl_public.h"
+#include "lv_types.h"
+#include "../draw/lv_draw_buf.h"
 
 /*********************
  *      DEFINES
@@ -76,7 +77,7 @@ static inline uint32_t lv_swap_bytes_32(uint32_t x)
  */
 static inline uint16_t lv_swap_bytes_16(uint16_t x)
 {
-    return (uint16_t)(x << 8) | (x >> 8);
+    return (x << 8) | (x >> 8);
 }
 
 /**********************
